OpenVINOQuantizer complete tests
